requirements.txt
llama-index
python-dotenv
pandas
langchain
streamlit
google-generativeai
langchain-google-genai
llama-index-llms-gemini
matplotlib
pandasai[google]